Reducing the number of default options set for the CLI
As these options are now specific to the CLI SAPI only - a number of default options could be considered unsuitable, or unnecessary.

For example, exposing PHP doesn't matter on the CLI, and limiting the script execution time may break long-running commands. These options make sense in a web-server context where the potential to cause damage through resource overloading for example are much higher.