Fixing annoying CI bug
The system-users role task to set its manifest information doesn't seem to work correctly in this role. Playing through the test playbook once gives a change event, as does the second, but subsequent plays are as expected (no change).

The system-users role itself does not have this bug, and unless this arises in other roles I am classing this as a bug. There is an easy workaround to get round this by simply playing the playbook a second time before checking if change events occurred.